+++ docs/changelog.rst
@@ -2,6 +2,19 @@
 CHANGELOG
 =========
 
+2.24.6 (2019-03-07)
+-------------------
+
+**Bug fixes**
+
+- When updating interventions, stake field is no more required
+- Fix duplicates in year filters in intervention module
+- Configurable blade code
+- Allow letters in blade number
+- Improve signage templates
+- Add "On signage/infrastructure" filter on intervention list
+
+
 2.24.5 (2019-03-06)
 -------------------